requirements.txt
praw
python-dateutil
nltk
tinydb
colorlog